backend/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.3
pydantic-settings==2.1.0
supabase==2.3.0
xai-sdk>=1.6.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
